Use proxy wiring for birth round shim
