+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/wiring/PlatformWiring.java
@@ -31,6 +31,7 @@ import com.swirlds.common.wiring.component.ComponentWiring;
 import com.swirlds.common.wiring.counters.BackpressureObjectCounter;
 import com.swirlds.common.wiring.counters.ObjectCounter;
 import com.swirlds.common.wiring.model.WiringModel;
+import com.swirlds.common.wiring.schedulers.builders.TaskSchedulerType;
 import com.swirlds.common.wiring.transformers.WireTransformer;
 import com.swirlds.common.wiring.wires.input.InputWire;
 import com.swirlds.common.wiring.wires.output.OutputWire;
@@ -76,7 +77,6 @@ import com.swirlds.platform.system.status.PlatformStatusManager;
 import com.swirlds.platform.system.status.actions.CatastrophicFailureAction;
 import com.swirlds.platform.util.HashLogger;
 import com.swirlds.platform.wiring.components.ApplicationTransactionPrehandlerWiring;
-import com.swirlds.platform.wiring.components.BirthRoundMigrationShimWiring;
 import com.swirlds.platform.wiring.components.ConsensusRoundHandlerWiring;
 import com.swirlds.platform.wiring.components.EventCreationManagerWiring;
 import com.swirlds.platform.wiring.components.EventDurabilityNexusWiring;
@@ -142,7 +142,7 @@ public class PlatformWiring implements Startable, Stoppable, Clearable {
     private final HashLoggerWiring hashLoggerWiring;
     private final LatestCompleteStateNotifierWiring latestCompleteStateNotifierWiring;
     private final PlatformCoordinator platformCoordinator;
-    private final BirthRoundMigrationShimWiring birthRoundMigrationShimWiring;
+    private final ComponentWiring<BirthRoundMigrationShim, GossipEvent> birthRoundMigrationShimWiring;
 
     /**
      * Constructor.
@@ -180,7 +180,13 @@ public class PlatformWiring implements Startable, Stoppable, Clearable {
                 .getConfigData(EventConfig.class)
                 .getAncientMode();
         if (ancientMode == AncientMode.BIRTH_ROUND_THRESHOLD) {
-            birthRoundMigrationShimWiring = BirthRoundMigrationShimWiring.create(model);
+            birthRoundMigrationShimWiring = new ComponentWiring<>(
+                    model,
+                    BirthRoundMigrationShim.class,
+                    model.schedulerBuilder("birthRoundMigrationShim")
+                            .withType(TaskSchedulerType.DIRECT_THREADSAFE)
+                            .build()
+                            .cast());
         } else {
             birthRoundMigrationShimWiring = null;
         }
@@ -282,9 +288,9 @@ public class PlatformWiring implements Startable, Stoppable, Clearable {
         final InputWire<GossipEvent> pipelineInputWire;
         if (birthRoundMigrationShimWiring != null) {
             birthRoundMigrationShimWiring
-                    .eventOutput()
+                    .getOutputWire()
                     .solderTo(eventHasherWiring.getInputWire(EventHasher::hashEvent));
-            pipelineInputWire = birthRoundMigrationShimWiring.eventInput();
+            pipelineInputWire = birthRoundMigrationShimWiring.getInputWire(BirthRoundMigrationShim::migrateEvent);
         } else {
             pipelineInputWire = eventHasherWiring.getInputWire(EventHasher::hashEvent);
         }
